Tests!
Implement a temporary measure as well that adds the module to your python path, letting you import to your heart's content. Also implement a unary - operator for Vector2D, because it makes life that much easier, and fixes a bug in util that caused multiplication to be done incorrectly if the Vector was on the right side of the equation.

Thanks to the new and improved Camera API, the code for the geometry test (now Tests/wobble.py) has been reduced from 48 lines to 18 lines (and if you don't include the manual module inclusion, 14 lines!) That's an improvement of 1-14/48 = 70.83%, pretty good if you ask me. (In fact, I might be able to reduce that by a line with some new updates I'm planning.