Add interpolation
And a button, "C", that will allow that interpolation
Comes with free object-is-in-screen detection for the Camera and some sort of sort of AABB! (For some reason I defined OBBs before AABBs. Don't ask me why...)
By popular request of self, Circles are now working ish! In the moveandwobble test, circles can be dragged with RMB, and squares with LMB. Also, arrows control camera movement, and you can scroll in and out.
Sorry for dumping documentation here, but this is a pretty temporary thing.